Refactor SmartMarketDataFetcher into delegated src/data source-fetch/metric-extraction/merge/gap-fill seams while preserving runtime and patch surfaces, and update ownership sensitive tests to follow the extracted modules.
