+++ tests/validators/test_metric_extractor.py
@@ -1,0 +1,15 @@
+"""Focused metric-extraction coverage for the red-flag validator."""
+
+from tests.validators.red_flag_validator_cases import (
+    TestDataBlockMarkerVariants,
+    TestDebtToEquityNormalization,
+    TestMetricExtraction,
+    TestSegmentOwnershipOCFFields,
+)
+
+__all__ = [
+    "TestMetricExtraction",
+    "TestDataBlockMarkerVariants",
+    "TestSegmentOwnershipOCFFields",
+    "TestDebtToEquityNormalization",
+]
