+++ tests/validators/test_red_flag_orchestrator.py
@@ -1,0 +1,13 @@
+"""Focused orchestration coverage for the red-flag validator node."""
+
+from tests.validators.red_flag_validator_cases import (
+    TestRedFlagIntegration,
+    TestRedFlagValidatorNode,
+    TestStrictValidatorNodeEscalation,
+)
+
+__all__ = [
+    "TestRedFlagValidatorNode",
+    "TestRedFlagIntegration",
+    "TestStrictValidatorNodeEscalation",
+]
