+++ tests/validators/test_supplemental_flags.py
@@ -1,0 +1,5 @@
+"""Focused supplemental-flag coverage for consultant-related validator logic."""
+
+from tests.validators.red_flag_validator_cases import TestConsultantConditionEnforcement
+
+__all__ = ["TestConsultantConditionEnforcement"]
